Add support for editing from bup
With these changes, the setupsheet can be printed and filled out in bup.
This allows using the same setupsheet that virtually all other clubs in the Bundesliga use.
It also simplifies the entry process, as the organizers don't have to select the names under time pressure, but can just copy the Xs.